+++ app/controllers/ActivityController.php
@@ -27,7 +27,7 @@
  *
  * Problem: Add more function to tradiccional admin.
  * @author $Author: Manuel Gil. $
- * @version $Revision: 0.2.3 $ $Date: 02/14/2021 $
+ * @version $Revision: 0.3.0 $ $Date: 02/15/2021 $
  * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  */
 
@@ -104,9 +104,9 @@ class ActivityController extends BaseController
 	 * <b>post: </b>AJAX request.
 	 *
 	 * @param string $module - the module name
-	 * @param null|string $courseid - the course id
+	 * @param string|null $courseid - the course id
 	 */
-	public function getListModule(string $module, $courseid = '')
+	public function getListModule(string $module, ?string $courseid = '')
 	{
 		// Imports Database.
 		global $DB;
@@ -167,6 +167,20 @@ class ActivityController extends BaseController
 			header_remove();
 			http_response_code(404);
 			header('HTTP/1.1 404 Not Found');
+			echo "	<head>
+						<link rel=\"stylesheet\" href=\"" . BASE_URL . "assets/css/slim.min.css\">
+						<script src=\"" . BASE_URL . "assets/js/slim.min.js\"></script>
+					</head>
+					<body>
+						<div class=\"page-error-wrapper\">
+							<div>
+								<h1 class=\"error-title\">404</h1>
+								<h5 class=\"tx-sm-24 tx-normal\">Oopps. The page you were looking for doesn't exist.</h5>
+								<p class=\"mg-b-50\">You may have mistyped the address or the page may have moved.</p>
+								<p class=\"mg-b-50\"><a href=\"" . BASE_URL . "\" class=\"btn btn-error\">Back to Home</a></p>
+							</div>
+						</div>
+					</body>";
 		}
 	}